pyproject.toml
[project]
name = "chase-bill-tracker"
version = "0.1.0"
description = "A tool to track Chase bills."
dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-generativeai",
    "requests",
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"